requirements.txt
asgiref==3.11.0
Django==4.2.26
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
sqlparse==0.5.4
